feat: extract building logic
